Rename tracking manager used for offloading
