+++ src/accel/CMakeLists.txt
@@ -34,8 +34,8 @@ list(APPEND SOURCES
 
 celeritas_polysource(ExceptionConverter)
 
-if(Geant4_VERSION VERSION_GREATER_EQUAL 11.0)
-  list(APPEND SOURCES TrackingManagerOffload.cc)
+if(Geant4_VERSION VERSION_GREATER_EQUAL 11.1)
+  list(APPEND SOURCES TrackingManager.cc)
 endif()
 
 if(CELERITAS_USE_HepMC3)
