+++ src/accel/TrackingManagerOffload.hh
@@ -6,41 +6,12 @@
 //---------------------------------------------------------------------------//
 #pragma once
 
-#include <G4VTrackingManager.hh>
+#include "TrackingManagerOffload.hh"
 
 namespace celeritas
 {
-class SharedParams;
-class LocalTransporter;
 
-//---------------------------------------------------------------------------//
-/*!
- * Offload tracks to Celeritas via the per-particle G4VTrackingManager
- * interface
- */
-class TrackingManagerOffload final : public G4VTrackingManager
-{
-  public:
-    // Construct with shared (across threads) params, and thread-local
-    // transporter.
-    TrackingManagerOffload(SharedParams const* params, LocalTransporter* local);
+// TODO: Remove in v0.7
+using TrackingManagerOffload [[deprecated]] = TrackingManager;
 
-    // Prepare cross-section tables for rebuild (e.g. if new materials have
-    // been defined).
-    void PreparePhysicsTable(G4ParticleDefinition const&) final;
-
-    // Rebuild physics cross-section tables (e.g. if new materials have been
-    // defined).
-    void BuildPhysicsTable(G4ParticleDefinition const&) final;
-
-    // Hand over passed track to this tracking manager.
-    void HandOverOneTrack(G4Track* aTrack) final;
-
-    // Complete processing of any buffered tracks.
-    void FlushEvent() final;
-
-  private:
-    SharedParams const* params_{nullptr};
-    LocalTransporter* transport_{nullptr};
-};
 }  // namespace celeritas
